cluster/swarm:fix small code

+++ cluster/swarm/database/system_config.go
@@ -17,9 +17,9 @@ type Configurations struct {
 	Registry
 	SSHDeliver
 
-	DockerPort int `db:"docker_port"`
-	PluginPort int `db:"plugin_port"`
-	Retry      int `db:"retry"`
+	DockerPort int   `db:"docker_port"`
+	PluginPort int   `db:"plugin_port"`
+	Retry      int64 `db:"retry"`
 }
 
 type SSHDeliver struct {
